refactor: extract AST processing module
- Create DprintMarkdownFormatter.AstProcessor module for AST operations
- Extract parse_elixir_source, collect_patches_for_doc_attributes, process_node_for_patches, create_patch_for_formatted_content functions
- Move all patch creation and string replacement logic to new module
- Reduce main module from 447 lines to 313 lines (~30% reduction)
- Maintain identical functionality and test coverage
